huge updatet 20
switch from interval to async/interval/timeout functions. 
This solve a host a response times as well as reducing code size.
Rewrote most of the logic systems, - major one is change the detection of playlist existence based on a scrollable content instead of != null, looking forward to the reduction in the bugs.